Edits and balancing
- Rebalanced Floran Set to be pre eye
- Edited Mime bomb sprite
- Bottom Feeder no longer spawns naturally
- New Spirit Trees, WIP Background, map background, edits to sky
- Cryolite now glows
- edited the Skull Stick to place anywhere
- Edited Brine Barrage spread, Bloomshroom collision detection, Tome of a Thousand Fangs balance and sfx

+++ Projectiles/Hostile/BloomshroomHostile.cs
@@ -31,7 +31,7 @@ namespace SpiritMod.Projectiles.Hostile
 		{
 			var list = Main.projectile.Where(x => x.Hitbox.Intersects(projectile.Hitbox));
 			foreach(var proj in list) {
-				if(projectile != proj && proj.friendly && proj.damage > 0) {
+				if(projectile != proj && proj.friendly && proj.damage > 0 && proj.active) {
 					projectile.Kill();
 				}
 			}
